misc. updates, mostly about portability and warnings removing some unnecessary #includes

+++ llt/random.h
@@ -7,8 +7,8 @@ double rand_double();
 float rand_float();
 double randn();
 void randomize();
-unsigned long genrand_int32();
-void init_genrand(unsigned long s);
+uint32_t genrand_int32();
+void init_genrand(uint32_t s);
 u_int64_t i64time();
 
 #endif
